restructure : large js code in independent files
