+++ py/prospect/scripts/__init__.py
@@ -1,0 +1,8 @@
+#
+# See top-level LICENSE.rst file for Copyright information
+#
+# -*- coding: utf-8 -*-
+
+# help with 2to3 support.
+from __future__ import absolute_import, division, print_function
+
